+++ CHANGELOG.md
@@ -2,6 +2,8 @@ CHANGELOG
 =========
 
 ## Unreleased
+- Retry renaming a temporary folder during plugin installation
+  [#3008](https://github.com/pulumi/pulumi/pull/3008)
 
 - Add support for additional Pulumi secrets providers using AWS KMS, Azure KeyVault, Google Cloud
   KMS and HashiCorp Vault.  These secrets providers can be configured at stack creation time using
@@ -12,6 +14,8 @@ CHANGELOG
   [#2994](https://github.com/pulumi/pulumi/pull/2994)
 
 - Add `Ouput.concat` to Python SDK [#3006](https://github.com/pulumi/pulumi/pull/3006)
+
+- Add `requireOutput` to `StackReference` [#3007](https://github.com/pulumi/pulumi/pull/3007)
 
 ### Compatibility
 
@@ -74,6 +78,7 @@ CHANGELOG
   by setting the `import` property in the resource options bag when instantiating a resource. In order to
   successfully import a resource, its desired configuration (i.e. its inputs) must not differ from its
   actual configuration (i.e. its state) as calculated by the resource's provider.
+- Better error message for missing npm on `pulumi new` (fixes [#1511](https://github.com/pulumi/pulumi/issues/1511))
 
 - Add the ability to pass a customTimeouts object from the providers across the engine
   to resource management. (fixes [#2655](https://github.com/pulumi/pulumi/issues/2655))